requirements.txt
fastapi
python-multipart
uvicorn
pillow
requests
jinja2
aiofiles
numpy
torch
torchvision
git+https://github.com/openai/CLIP.git
google-generativeai
chromadb
scikit-learn
tensorflow==2.13
opencv-python
pandas
python-dotenv
